requirements.txt
Flask==2.1.0
google-cloud-datastore==2.7.1
flake8
black
isort
mypy
python-dateutil
google-cloud-tasks
requests
